add profiler to plotter helper and convert profiling plots into funcs
